final fixes to handle errors in printf + refactoring

+++ printf/ft_printf.h
@@ -17,7 +17,7 @@
 
 int		ft_printf(const char *str, ...);
 int		ft_printchar(int c);
-int		ft_printhex(unsigned long arg, char c);
+int		ft_printhex(unsigned int arg, char c);
 int		ft_printnum(int n, char c);
 int		ft_printstr(char *str);
 int		ft_printpointer(void *arg);
